create new RequestReader class and update tests
