+++ src/main/java/server/ServerSocketWrapper.java
@@ -2,14 +2,8 @@ package server;
 
 import java.io.IOException;
 import java.net.Socket;
-import java.util.HashMap;
-import server.constants.Format;
-import server.constants.Header;
-import server.constants.Method;
-import server.constants.Path;
 import server.request.Request;
-import server.request.RequestBuilder;
-import server.request.RequestParser;
+import server.request.RequestReader;
 import server.response.Response;
 import server.response.ResponseFormatter;
 
@@ -23,48 +17,7 @@ public class ServerSocketWrapper implements ServerSocketInterface {
   }
 
   public Request getRequest() throws IOException {
-    String rawRequest = getRequestHead();
-
-    if (!rawRequest.isBlank()) {
-      RequestParser parser = new RequestParser(rawRequest);
-      Method method = parser.method();
-      Path path = parser.path();
-      HashMap<Header, String> headers = parser.headers();
-
-      String body = null;
-      if (headers.containsKey(Header.CONTENT_LENGTH)) {
-        int bytes = Integer.parseInt(headers.get(Header.CONTENT_LENGTH));
-        body = getRequestBody(bytes);
-      }
-
-      return new RequestBuilder()
-              .setMethod(method)
-              .setPath(path)
-              .setHeaders(headers)
-              .setBody(body)
-              .build();
-    }
-    return null;
-  }
-
-  private String getRequestHead() throws IOException {
-    StringBuilder request = new StringBuilder();
-    String line;
-
-    while ((line = socketIo.readLine()) != null) {
-      request.append(line).append(Format.BREAK);
-
-      if (line.isBlank()) {
-        break;
-      }
-    }
-
-    return request.toString();
-  }
-
-  private String getRequestBody(int length) throws IOException {
-    return socketIo.readBytes(length);
-
+    return new RequestReader(socketIo).readRequest();
   }
 
   public void sendResponse(Response response) {
